refactor(packages): move non-internal packages to /pkg
