+++ internal/validate_paths/validate_paths.go
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package validate
+package validate_paths
 
 import (
 	"errors"
