+++ pkg/parse_hal/parse_hal.go
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package parse
+package parse_hal
 
 import (
 	"github.com/ghodss/yaml"
