+++ pkg/validate_hal/validate_hal.go
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package validate
+package validate_hal
 
 import (
 	"errors"
